Fixed some bugs in magma circuits
Notified Magma/Mantle developers of a small issue. This was fixed in
the latest version of Mantle.

The verilog backend of Magma has not been maintained as well as the
lattice or coreir backend. Magma developers recommended producing
coreir then using the coreir command line to produce verilog. This
compiles, but I have not checked the circuits for correctness yet.